refactor: split tablePositioning
Split tableRuntime/tablePositioning.ts into clear layers so DOM/event positioning can depend on getResolvedActiveCell() without creating a circular dependency
